Review Dependabot alerts for `net6.0` and `net7.0` target frameworks and bump all packages (#2220)
* Fix CS1574 warning

* IDE0060: Remove unused parameter 'queryString' if it is not part of a shipped public API

* Warning NETSDK1138: The target framework 'net7.0' is out of support and will not receive security updates in the future. Please refer to https://aka.ms/dotnet-core-support for more information about the support policy.
Remove target framework 'net7.0' from testing projects.

* IDE0079: Remove unnecessary suppression SA1649

* Bump all packages

* Fix issues of code review by @ggnaegi
