+++ src/Ocelot.Provider.Consul/Ocelot.Provider.Consul.csproj
@@ -30,7 +30,7 @@
     <ProjectReference Include="..\Ocelot\Ocelot.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Consul" Version="1.7.14.4" />
+    <PackageReference Include="Consul" Version="1.7.14.5" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.556">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
