+++ src/Ocelot/Responses/OkResponse.cs
@@ -2,7 +2,10 @@ namespace Ocelot.Responses;
 
 public class OkResponse : Response
 {
-    public OkResponse()
-    {
-    }
+    public OkResponse() { }
+}
+
+public class OkResponse<T> : Response<T>
+{
+    public OkResponse(T data) : base(data) { }
 }
