+++ src/Ocelot/Responses/Response.cs
@@ -18,3 +18,16 @@ public abstract class Response
 
     public bool IsError => Errors.Count > 0;
 }
+
+public abstract class Response<T> : Response
+{
+    protected Response(T data)
+    {
+        Data = data;
+    }
+
+    protected Response(List<Error> errors) : base(errors)
+    { }
+
+    public T Data { get; }
+}
